Рефакторинг пакета reporter и перевод на рельсы спринга
