+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/reporters/TSLintReporter.java
@@ -1,0 +1,65 @@
+/*
+ * This file is a part of BSL Language Server.
+ *
+ * Copyright © 2018-2020
+ * Alexey Sosnoviy <labotamy@gmail.com>, Nikita Gryzlov <nixel2007@gmail.com> and contributors
+ *
+ * SPDX-License-Identifier: LGPL-3.0-or-later
+ *
+ * BSL Language Server is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3.0 of the License, or (at your option) any later version.
+ *
+ * BSL Language Server is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with BSL Language Server.
+ */
+package com.github._1c_syntax.bsl.languageserver.reporters;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.github._1c_syntax.bsl.languageserver.reporters.data.AnalysisInfo;
+import com.github._1c_syntax.bsl.languageserver.reporters.data.FileInfo;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import org.eclipse.lsp4j.Diagnostic;
+import org.springframework.stereotype.Component;
+
+import java.io.File;
+import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.List;
+
+@Slf4j
+@Component
+public class TSLintReporter implements DiagnosticReporter {
+
+  @Override
+  public String key() {
+    return "tslint";
+  }
+
+  @Override
+  @SneakyThrows
+  public void report(AnalysisInfo analysisInfo, Path outputDir) {
+    List<TSLintReportEntry> tsLintReport = new ArrayList<>();
+    for (FileInfo fileInfo : analysisInfo.getFileinfos()) {
+      for (Diagnostic diagnostic : fileInfo.getDiagnostics()) {
+        TSLintReportEntry entry = new TSLintReportEntry(fileInfo.getPath().toString(), diagnostic);
+        tsLintReport.add(entry);
+      }
+    }
+
+    ObjectMapper mapper = new ObjectMapper();
+    mapper.enable(SerializationFeature.INDENT_OUTPUT);
+
+    File reportFile = new File(outputDir.toFile(), "./bsl-tslint.json");
+    mapper.writeValue(reportFile, tsLintReport);
+    LOGGER.info("TSLint report saved to {}", reportFile.getAbsolutePath());
+  }
+}
