+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/reporters/databind/DiagnosticCodeDeserializer.java
@@ -19,39 +19,28 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with BSL Language Server.
  */
-package com.github._1c_syntax.bsl.languageserver.diagnostics.reporter;
+package com.github._1c_syntax.bsl.languageserver.reporters.databind;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github._1c_syntax.bsl.languageserver.diagnostics.databind.AnalysisInfoObjectMapper;
-import lombok.extern.slf4j.Slf4j;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+import com.github._1c_syntax.bsl.languageserver.diagnostics.metadata.DiagnosticCode;
+import org.eclipse.lsp4j.jsonrpc.messages.Either;
 
-import java.io.File;
 import java.io.IOException;
-import java.nio.file.Path;
 
-@Slf4j
-public class JsonReporter extends AbstractDiagnosticReporter {
-
-  public static final String KEY = "json";
-
-  public JsonReporter() {
-    super();
-  }
-
-  public JsonReporter(Path outputDir) {
-    super(outputDir);
-  }
+/**
+ * Десериализатор для {@link Either}, выступающего в роли хранилища кода диагностики.
+ * См. {@link DiagnosticCode}
+ */
+public class DiagnosticCodeDeserializer extends JsonDeserializer<Either<String, Number>> {
 
   @Override
-  public void report(AnalysisInfo analysisInfo) {
-    ObjectMapper mapper = new AnalysisInfoObjectMapper();
-
-    try {
-      File reportFile = new File(outputDir.toFile(), "./bsl-json.json");
-      mapper.writeValue(reportFile, analysisInfo);
-      LOGGER.info("JSON report saved to {}", reportFile.getAbsolutePath());
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
+  public Either<String, Number> deserialize(
+    JsonParser p,
+    DeserializationContext ctxt
+  ) throws IOException {
+    return Either.forLeft(p.getValueAsString());
   }
+
 }
